call new spine function for text mention notification as Html
